+++ starkinfra/individualaccountrequest/__init__.py
@@ -1,3 +1,4 @@
 from .__individualaccountrequest import create, get, query, page, update
+from .__address import Address
 from .log.__log import Log
 from . import log